requirements.txt
SQLAlchemy==2.0.25
python-dotenv==1.0.0
async-lru==2.0.4
asyncpg==0.29.0
discord.py==2.3.2
greenlet
openai==1.12.0
pytz==2024.1

#below are for testing
pytest-asyncio
pytest
sentry-sdk==2.7.1